clap and yoyoke + dice improve and imput error

+++ bot.py
@@ -22,13 +22,21 @@ async def clap(ctx):
     await ctx.send(f'{ctx.author.mention} failed at clapping!' if claps == 0 else f'{ctx.author.mention} clapped! {":clap:" * claps}')
 
 @bot.command(name='roll_dice', help='Simulates rolling dice.')
-async def roll(ctx, number_of_dice: int, number_of_sides: int):
-    dice = [
-        str(random.choice(range(1, number_of_sides + 1)))
-        for _ in range(number_of_dice)
-    ]
-    await ctx.message.delete()
-    await ctx.send(', '.join(dice))
+async def roll(ctx, number_of_dice: int, number_of_sides: str):
+    if not number_of_sides in ["d4", "d6", "d8", "d10", "d12", "d20", "d100"]:
+        raise commands.errors.UserInputError
+    elif number_of_sices > 500:
+        await ctx.send("Too many dice (max = 500).")
+    else:
+        dice = [
+            str(random.choice(range(1, int(number_of_sides[1:]) + 1)))
+            for _ in range(number_of_dice)
+        ]
+        await ctx.message.delete()
+        if number_of_dice > 100:
+            await ctx.send(f'Total {number_of_dice}: {sum([int(x) for x in dice])}\n_Collapsed to reduce spam._')
+        else:
+            await ctx.send(', '.join(dice))
 
 @bot.command(name='info', help='Get info about the bot.')
 async def info(ctx):
@@ -42,6 +50,13 @@ async def info(ctx):
     await ctx.send(file=file, embed=embed)
 
 @bot.event
+async def on_message(message):
+    if message.author == bot.user:
+        return
+    if 'yoyoke' in message.content.split(" "):
+        await message.channel.send("HAHA YOYOKE!")
+
+@bot.event
 async def on_error(event, *args, **kwargs):
     if event == 'on_message':
         await args[0].author.create_dm()
@@ -53,5 +68,7 @@ async def on_error(event, *args, **kwargs):
 async def on_command_error(ctx, error):
     if isinstance(error, commands.errors.CheckFailure):
         await ctx.send('You do not have the correct role for this command.')
+    elif isinstance(error, commands.errors.UserInputError):
+        await ctx.send('Argument error.')
 
 bot.run(TOKEN)
